+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/ProfitLossCriterionTest.java
@@ -39,6 +39,28 @@ public class ProfitLossCriterionTest  extends AbstractCriterionTest {
         assertNumEquals(-250 - 1500, profit.calculate(series, tradingRecord));
     }
 
+    @Test
+    public void calculateShortOnlyWithGainTrades() {
+        MockTimeSeries series = new MockTimeSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        TradingRecord tradingRecord = new BaseTradingRecord(
+                Order.sellAt(0,series, series.numOf(50)), Order.buyAt(2,series, series.numOf(50)),
+                Order.sellAt(3,series, series.numOf(50)), Order.buyAt(5,series, series.numOf(50)));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals(-(500 + 250), profit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateShortOnlyWithLossTrades() {
+        MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        TradingRecord tradingRecord = new BaseTradingRecord(
+                Order.sellAt(0,series, series.numOf(50)), Order.buyAt(1,series, series.numOf(50)),
+                Order.sellAt(2,series, series.numOf(50)), Order.buyAt(5,series, series.numOf(50)));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals(250 + 1500, profit.calculate(series, tradingRecord));
+    }
+
 
     @Test
     public void betterThan() {