+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Fixed `java.lang.NullPointerException` in**: `AverageProfitableTradesCriterion.calculate(TimeSeries, Trade)` for opened trade.
 - **StopGainRule**: now correctly handles stops for sell orders
 - **StopLossRule**: now correctly handles stops for sell orders
+- **ProfitLossCriterion*: fixed to work properly for short trades
 
 ### Changed